# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ArtFi API
  version: 0.4.0
  description: Stage 4 Base Sepolia preview contract. Wallet intents never hold private keys; authenticated indexer writes are durable and replay-safe.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/assets:
    get:
      operationId: listAssets
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 120 } }
        - { name: status, in: query, schema: { type: string, maxLength: 40 } }
        - {
            name: page,
            in: query,
            schema: { type: integer, minimum: 1, default: 1 },
          }
        - {
            name: pageSize,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100, default: 20 },
          }
      responses:
        "200":
          description: Asset catalog
          content:
            application/json:
              schema:
                type: object
                required: [data, total, page, pageSize]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Asset"
                  total:
                    type: integer
                    minimum: 0
                  page: { type: integer, minimum: 1 }
                  pageSize: { type: integer, minimum: 1, maximum: 100 }
  /v1/assets/{slug}:
    get:
      operationId: getAsset
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Asset record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/nfts:
    get:
      operationId: listMintedNFTs
      description: Runtime Base Sepolia mint catalog projected from canonical chain events.
      parameters:
        - {
            name: page,
            in: query,
            schema: { type: integer, minimum: 1, default: 1 },
          }
        - {
            name: pageSize,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100, default: 20 },
          }
      responses:
        "200":
          description: Indexed ERC-721 and ERC-1155 mint records
          content:
            application/json:
              schema:
                type: object
                required: [data, total, page, pageSize, chainId, runtime]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MintedNFT"
                  total: { type: integer, minimum: 0 }
                  page: { type: integer, minimum: 1 }
                  pageSize: { type: integer, minimum: 1, maximum: 100 }
                  chainId: { type: integer, const: 84532 }
                  runtime: { type: boolean, const: true }
  /v1/projects:
    get:
      operationId: listProjects
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 120 } }
        - {
            name: page,
            in: query,
            schema: { type: integer, minimum: 1, default: 1 },
          }
        - {
            name: pageSize,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100, default: 20 },
          }
      responses:
        "200":
          description: Curated projects
          content:
            application/json:
              schema:
                type: object
                required: [data, total, page, pageSize]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
                  total:
                    type: integer
                    minimum: 0
                  page: { type: integer, minimum: 1 }
                  pageSize: { type: integer, minimum: 1, maximum: 100 }
  /v1/portfolio/{address}:
    get:
      operationId: getPortfolio
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
            pattern: "^0x[0-9a-fA-F]{40}$"
      responses:
        "200":
          description: Read-only public address portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
        "400":
          $ref: "#/components/responses/Problem"
  /v1/indexer/events:
    post:
      operationId: ingestChainEvent
      parameters:
        - name: X-Indexer-Key
          in: header
          required: true
          schema: { type: string, minLength: 16 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChainEventRequest"
      responses:
        "201": { description: Canonical event created }
        "200":
          {
            description: Replay,
            confirmation update,
            removal,
            or reorg replacement recorded,
          }
        "401": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /v1/indexer/market-events:
    post:
      operationId: ingestExternalMarketEvent
      description: Credential-gated ingestion for normalized, read-only events from approved marketplaces.
      parameters:
        - name: X-Indexer-Key
          in: header
          required: true
          schema: { type: string, minLength: 16 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExternalMarketEventRequest"
      responses:
        "201": { description: External marketplace event mirrored }
        "200": { description: Idempotent replay accepted }
        "401": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /v1/market/activity:
    get:
      operationId: getExternalMarketActivity
      description: Read-only activity mirrored from approved external marketplaces; execution remains external.
      parameters:
        - name: source
          in: query
          schema: { type: string, default: opensea }
        - name: contract
          in: query
          schema: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: Attributed external marketplace activity
          content:
            application/json:
              schema:
                type: object
                required: [data, schemaVersion, source, execution, custody]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ExternalMarketActivity"
                  schemaVersion: { type: string, const: "1" }
                  source: { type: string }
                  execution: { type: string, const: external-deeplink-only }
                  custody: { type: boolean, const: false }
        "400": { $ref: "#/components/responses/Problem" }
  /v1/market/assets:
    get:
      operationId: getExternalMarketAssets
      description: Paginated, deduplicated NFT catalog derived only from attributed external-market runtime events.
      parameters:
        - name: source
          in: query
          schema: { type: string, default: opensea }
        - name: contract
          in: query
          schema: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Complete runtime-backed page of mirrored NFTs
          content:
            application/json:
              schema:
                type: object
                required:
                  [data, total, page, pageSize, schemaVersion, source, runtime]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ExternalMarketAsset" }
                  total: { type: integer, minimum: 0 }
                  page: { type: integer, minimum: 1 }
                  pageSize: { type: integer, minimum: 1, maximum: 100 }
                  schemaVersion: { type: string, const: "1" }
                  source: { type: string }
                  runtime: { type: boolean, const: true }
        "400": { $ref: "#/components/responses/Problem" }
  /v1/market/intents:
    post:
      operationId: createExternalMarketIntent
      description: Build an unsigned, same-chain OpenSea fulfillment plan. The external wallet and OpenSea/Seaport retain signing and execution.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 16, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExternalMarketIntentRequest" }
      responses:
        "201":
          description: Safe unsigned transaction plan prepared for external-wallet review
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExternalMarketIntent" }
        "200":
          description: Idempotent replay
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExternalMarketIntent" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /v1/market/intents/{intentID}:
    get:
      operationId: getExternalMarketIntent
      parameters:
        - name: intentID
          in: path
          required: true
          schema: { type: string, pattern: "^[0-9a-f]{32}$" }
      responses:
        "200":
          description: Current state reconciled from wallet submission and external-market events
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExternalMarketIntent" }
        "404": { $ref: "#/components/responses/Problem" }
  /v1/market/intents/{intentID}/submission:
    post:
      operationId: recordExternalMarketSubmission
      parameters:
        - name: intentID
          in: path
          required: true
          schema: { type: string, pattern: "^[0-9a-f]{32}$" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [transactionHash]
              properties:
                transactionHash:
                  { type: string, pattern: "^0x[0-9a-fA-F]{64}$" }
      responses:
        "200":
          description: External-wallet submission recorded for authoritative event reconciliation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExternalMarketIntent" }
        "409": { $ref: "#/components/responses/Problem" }
...
  /v1/governance/proposals:
    get:
      operationId: getGovernanceProposals
      description: Durable chain-indexed governance proposals. No fixture fallback is permitted.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: Indexed governance proposal records
          content:
            application/json:
              schema:
                type: object
                required: [data, membershipAuthority, schemaVersion]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/GovernanceProposal" }
                  membershipAuthority:
                    { type: string, const: onchain-rwa-token-snapshot }
                  schemaVersion: { type: string, const: "1" }
        "503": { $ref: "#/components/responses/Problem" }
  /v1/uploads/intents:
    post:
      operationId: createUploadIntent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadIntentRequest"
      responses:
        "201":
          description: Short-lived digest-bound API upload intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadIntent"
        "422":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /v1/uploads/{uploadID}:
    put:
      operationId: uploadObject
      parameters:
        - name: uploadID
          in: path
          required: true
          schema: { type: string, pattern: "^[0-9a-f]{32}$" }
        - name: Content-SHA256
          in: header
          required: true
          schema: { type: string, pattern: "^[0-9a-f]{64}$" }
      requestBody:
        required: true
        content:
          image/png: { schema: { type: string, format: binary } }
          image/jpeg: { schema: { type: string, format: binary } }
          image/webp: { schema: { type: string, format: binary } }
      responses:
        "204": { description: Object digest verified and persisted }
        "400":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
  /v1/rwa/intents:
    post:
      operationId: createMintIntent
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 16, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MintIntentRequest"
      responses:
        "201":
          description: New immutable metadata and contract-call intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MintIntent"
        "200":
          description: Idempotent replay of an existing intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MintIntent"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
  /v1/rwa/intents/{intentID}:
    get:
      operationId: getMintIntent
      parameters:
        - $ref: "#/components/parameters/IntentID"
      responses:
        "200":
          description: Mint transaction lifecycle snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MintIntent"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/rwa/intents/{intentID}/submission:
    post:
      operationId: recordMintSubmission
      parameters:
        - $ref: "#/components/parameters/IntentID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transactionHash]
              properties:
                transactionHash:
                  { type: string, pattern: "^0x[0-9a-fA-F]{64}$" }
      responses:
        "200":
          description: Submitted transaction attached idempotently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MintIntent"
        "409":
          $ref: "#/components/responses/Problem"
  /v1/rwa/discovery-checks:
    post:
      operationId: checkOpenSeaDiscovery
      description: Operator-authenticated, server-side OpenSea NFT discovery check. The API key and raw upstream payload remain private; the result and payload digest are recorded durably.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenSeaDiscoveryRequest"
      responses:
        "200":
          description: Discovery evidence recorded, including explicit not-found or unsupported-chain outcomes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenSeaDiscoveryResult"
        "401": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /v1/vault/intents:
    post:
      operationId: createVaultIntent
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 16, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultIntentRequest"
      responses:
        "201":
          description: New VaultFactory transaction intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultIntent"
        "200":
          description: Idempotent replay
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultIntent"
        "409":
          $ref: "#/components/responses/Problem"
  /v1/vault/intents/{intentID}:
    get:
      operationId: getVaultIntent
      parameters:
        - $ref: "#/components/parameters/IntentID"
      responses:
        "200":
          description: Vault transaction lifecycle snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultIntent"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/vault/intents/{intentID}/submission:
    post:
      operationId: recordVaultSubmission
      parameters:
        - $ref: "#/components/parameters/IntentID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transactionHash]
              properties:
                transactionHash:
                  { type: string, pattern: "^0x[0-9a-fA-F]{64}$" }
      responses:
        "200":
          description: Submitted VaultFactory transaction attached idempotently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultIntent"
components:
...
  responses:
    Problem:
      description: RFC 9457-style request failure
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Asset:
      type: object
      required:
        [
          slug,
          title,
          artist,
          year,
          medium,
          location,
          valuationUsd,
          fractionPriceUsd,
          totalFractions,
          availableFractions,
          status,
          provenance,
        ]
      properties:
        slug: { type: string }
        title: { type: string }
        artist: { type: string }
        year: { type: integer }
        medium: { type: string }
        location: { type: string }
        valuationUsd: { type: integer, minimum: 1 }
        fractionPriceUsd: { type: integer, minimum: 1 }
        totalFractions: { type: integer, minimum: 1 }
        availableFractions: { type: integer, minimum: 0 }
        status: { type: string }
        provenance:
          type: array
          items: { type: string }
    MintedNFT:
      type: object
      required:
        [
          standard,
          collectionAddress,
          tokenId,
          transactionHash,
          blockNumber,
          observedAt,
        ]
      properties:
        standard: { type: string, enum: [ERC-721, ERC-1155] }
        collectionAddress: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        tokenId: { type: string, pattern: "^(0|[1-9][0-9]{0,77})$" }
        holderAddress: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        metadataUri: { type: string, format: uri, pattern: "^https://" }
        transactionHash: { type: string, pattern: "^0x[0-9a-fA-F]{64}$" }
        blockNumber: { type: integer, minimum: 0 }
        observedAt: { type: string, format: date-time }
    Project:
      type: object
      required: [slug, name, curator, location, assetSlugs, description]
      properties:
        slug: { type: string }
        name: { type: string }
        curator: { type: string }
        location: { type: string }
        assetSlugs:
          type: array
          items: { type: string }
        description: { type: string }
    Portfolio:
      type: object
      required:
        [
          address,
          chainId,
          network,
          positions,
          transactions,
          offers,
          notifications,
        ]
      properties:
        address: { type: string }
        chainId: { type: integer, const: 84532 }
        network: { type: string, const: base-sepolia }
        positions:
          type: array
          items:
            type: object
            required: [assetToken, symbol, balance, updatedAt]
            properties:
              assetToken: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
              symbol: { type: string, maxLength: 12 }
              balance: { type: string, pattern: "^(0|[1-9][0-9]{0,77})$" }
              updatedAt: { type: string, format: date-time }
        transactions:
          type: array
          items: { type: object }
        offers:
          type: array
          items: { type: object }
        notifications:
          type: array
          items: { type: object }
    ChainEventRequest:
      type: object
      additionalProperties: false
      required:
        [
          chainId,
          transactionHash,
          logIndex,
          blockNumber,
          blockHash,
          contractAddress,
          eventName,
          payload,
          removed,
          confirmations,
        ]
      properties:
        chainId: { type: integer, const: 84532 }
        transactionHash: { type: string, pattern: "^0x[0-9a-fA-F]{64}$" }
        logIndex: { type: integer, minimum: 0 }
        blockNumber: { type: integer, minimum: 0 }
        blockHash: { type: string, pattern: "^0x[0-9a-fA-F]{64}$" }
        contractAddress: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        eventName: { type: string, minLength: 2, maxLength: 80 }
        payload: { type: object }
        removed: { type: boolean }
        confirmations: { type: integer, minimum: 0 }
    ExternalMarketEventRequest:
      type: object
      additionalProperties: false
      required:
        [
          schemaVersion,
          source,
          eventType,
          eventFamily,
          entityKey,
          version,
          chain,
          eventTimestamp,
          payload,
        ]
      properties:
        schemaVersion: { type: string, const: "1" }
        source: { type: string, pattern: "^[a-z][a-z0-9-]{1,31}$" }
        eventType: { type: string, maxLength: 48 }
        eventFamily: { type: string, enum: [order, sale, transfer, metadata] }
        entityKey: { type: string, maxLength: 196 }
        version: { type: integer, minimum: 1 }
        chain: { type: string, pattern: "^[a-z][a-z0-9-]{1,31}$" }
        collectionSlug: { type: string, maxLength: 160 }
        orderHash: { type: string, pattern: "^0x[0-9a-fA-F]{64}$" }
        transactionHash: { type: string, pattern: "^0x[0-9a-fA-F]{64}$" }
        contractAddress: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        tokenId: { type: string, pattern: "^(0|[1-9][0-9]{0,77})$" }
        makerAddress: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        price: { type: string, pattern: "^(0|[1-9][0-9]{0,77})$" }
        paymentTokenAddress: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        paymentSymbol: { type: string, maxLength: 16 }
        marketplaceUrl: { type: string, format: uri }
        eventTimestamp: { type: string, format: date-time }
        payload: { type: object }
    ExternalMarketActivity:
      allOf:
        - $ref: "#/components/schemas/ExternalMarketEventRequest"
        - type: object
          required: [eventId]
          properties:
            eventId: { type: string, pattern: "^[0-9a-f]{64}$" }
...
    ExternalMarketIntentRequest:
      type: object
      additionalProperties: false
      required: [source, action, chain, orderHash, walletAddress]
      properties:
        source: { type: string, const: opensea }
        action: { type: string, const: fulfill-listing }
        chain: { type: string }
        orderHash: { type: string, pattern: "^0x[0-9a-fA-F]{64}$" }
        walletAddress: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
...